parser-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-pptx>=1.0.2
python-multipart>=0.0.12
python-dotenv>=1.0.1
supabase>=2.10.0
